refactor(ai): isolate SQLite credential store
